games/dino: flatten collision math, byte-sized boxes, drop time()
Replace the struct-by-value createAdjustedCollisionBox/boxCompare pair
with one boxesIntersect() over plain ints; collision box fields shrink
to signed char (largest value is 58). Drop the trex.msPerFrame field
duplicating currentAnimFrames->msPerFrame, fold the repeated jump-key
test into isJumpKey(), seed rand from _ksys_get_ns_count instead of
time(NULL), and store minSpeed/frameRate as float (8.5 is exact in
float; the integer timer threshold is unchanged).

.text 12732 -> 12317, .data -280. Verified: 20M randomized collision
states match the old code exactly; the 20000-frame horizon simulation
stays byte-identical.

+++ programs/games/dino/obstacle.h
@@ -36,11 +36,11 @@ typedef struct {
     int yPosArr[3]; // used if yPos is -1
     int multipleSpeed;
     int minGap;
-    double minSpeed; // 8.5 for the pterodactyl, must not truncate
+    float minSpeed; // 8.5 for the pterodactyl: exact in float, must not truncate
     int collisionBoxesCount;
     CollisionBox collisionBoxes[5];
     int numFrames;
-    double frameRate;
+    float frameRate;
     double speedOffset;
 } ObstacleTypeConfig;
